Move ELASPIC CLI to a separate folder.
